+++ banyand/internal/cmd/liaison.go
@@ -24,6 +24,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/apache/skywalking-banyandb/api/common"
+	"github.com/apache/skywalking-banyandb/banyand/dquery"
 	"github.com/apache/skywalking-banyandb/banyand/liaison/grpc"
 	"github.com/apache/skywalking-banyandb/banyand/liaison/http"
 	"github.com/apache/skywalking-banyandb/banyand/metadata"
@@ -49,10 +50,15 @@ func newLiaisonCmd() *cobra.Command {
 	profSvc := observability.NewProfService()
 	metricSvc := observability.NewMetricService()
 	httpServer := http.NewServer()
+	dQuery, err := dquery.NewService(metaSvc, pipeline)
+	if err != nil {
+		l.Fatal().Err(err).Msg("failed to initiate distributed query service")
+	}
 
 	units := []run.Unit{
 		new(signal.Handler),
 		pipeline,
+		dQuery,
 		grpcServer,
 		httpServer,
 		profSvc,