+++ pkg/query/executor/interface.go
@@ -19,12 +19,15 @@
 package executor
 
 import (
+	"context"
+
 	"github.com/apache/skywalking-banyandb/api/common"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	measurev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/measure/v1"
 	modelv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/model/v1"
 	streamv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/stream/v1"
 	"github.com/apache/skywalking-banyandb/banyand/tsdb"
+	"github.com/apache/skywalking-banyandb/pkg/bus"
 )
 
 // ExecutionContext allows retrieving data from tsdb.
@@ -40,9 +43,24 @@ type StreamExecutionContext interface {
 	ParseElementID(item tsdb.Item) (string, error)
 }
 
+// StreamExecutionContextKey is the key of stream execution context in context.Context.
+type StreamExecutionContextKey struct{}
+
+var streamExecutionContextKeyInstance = StreamExecutionContextKey{}
+
+// WithStreamExecutionContext returns a new context with stream execution context.
+func WithStreamExecutionContext(ctx context.Context, ec StreamExecutionContext) context.Context {
+	return context.WithValue(ctx, streamExecutionContextKeyInstance, ec)
+}
+
+// FromStreamExecutionContext returns the stream execution context from context.Context.
+func FromStreamExecutionContext(ctx context.Context) StreamExecutionContext {
+	return ctx.Value(streamExecutionContextKeyInstance).(StreamExecutionContext)
+}
+
 // StreamExecutable allows querying in the stream schema.
 type StreamExecutable interface {
-	Execute(StreamExecutionContext) ([]*streamv1.Element, error)
+	Execute(context.Context) ([]*streamv1.Element, error)
 }
 
 // MeasureExecutionContext allows retrieving data through the measure module.
@@ -50,6 +68,21 @@ type MeasureExecutionContext interface {
 	ExecutionContext
 	ParseField(name string, item tsdb.Item) (*measurev1.DataPoint_Field, error)
 	CompanionShards(metadata *commonv1.Metadata) ([]tsdb.Shard, error)
+}
+
+// MeasureExecutionContextKey is the key of measure execution context in context.Context.
+type MeasureExecutionContextKey struct{}
+
+var measureExecutionContextKeyInstance = MeasureExecutionContextKey{}
+
+// WithMeasureExecutionContext returns a new context with measure execution context.
+func WithMeasureExecutionContext(ctx context.Context, ec MeasureExecutionContext) context.Context {
+	return context.WithValue(ctx, measureExecutionContextKeyInstance, ec)
+}
+
+// FromMeasureExecutionContext returns the measure execution context from context.Context.
+func FromMeasureExecutionContext(ctx context.Context) MeasureExecutionContext {
+	return ctx.Value(measureExecutionContextKeyInstance).(MeasureExecutionContext)
 }
 
 // MIterator allows iterating in a measure data set.
@@ -63,5 +96,26 @@ type MIterator interface {
 
 // MeasureExecutable allows querying in the measure schema.
 type MeasureExecutable interface {
-	Execute(MeasureExecutionContext) (MIterator, error)
+	Execute(context.Context) (MIterator, error)
+}
+
+// DistributedExecutionContext allows retrieving data through the distributed module.
+type DistributedExecutionContext interface {
+	bus.Broadcaster
+	TimeRange() *modelv1.TimeRange
+}
+
+// DistributedExecutionContextKey is the key of distributed execution context in context.Context.
+type DistributedExecutionContextKey struct{}
+
+var distributedExecutionContextKeyInstance = DistributedExecutionContextKey{}
+
+// WithDistributedExecutionContext returns a new context with distributed execution context.
+func WithDistributedExecutionContext(ctx context.Context, ec DistributedExecutionContext) context.Context {
+	return context.WithValue(ctx, distributedExecutionContextKeyInstance, ec)
+}
+
+// FromDistributedExecutionContext returns the distributed execution context from context.Context.
+func FromDistributedExecutionContext(ctx context.Context) DistributedExecutionContext {
+	return ctx.Value(distributedExecutionContextKeyInstance).(DistributedExecutionContext)
 }