+++ pkg/query/logical/measure/measure_plan_indexscan_local.go
@@ -20,6 +20,7 @@ package measure
 import (
 	"context"
 	"fmt"
+	"github.com/apache/skywalking-banyandb/pkg/iter/sort"
 	"io"
 	"time"
 
@@ -73,7 +74,7 @@ func (uis *unresolvedIndexScan) Analyze(s logical.Schema) (logical.Plan, error) 
 
 	entityList := s.EntityList()
 	entityMap, entity := uis.locateEntity(entityList)
-	filter, entities, err := logical.BuildLocalFilter(uis.criteria, s, entityMap, entity)
+	filter, entities, err := logical.BuildLocalFilter(uis.criteria, s, entityMap, entity, true)
 	if err != nil {
 		return nil, err
 	}
@@ -142,26 +143,28 @@ func (i *localIndexScan) Sort(order *logical.OrderBy) {
 	i.order = order
 }
 
-func (i *localIndexScan) Execute(ec executor.MeasureExecutionContext) (mit executor.MIterator, err error) {
+func (i *localIndexScan) Execute(ctx context.Context) (mit executor.MIterator, err error) {
 	if i.isTopN {
-		return i.topNExecute(ec)
+		return i.topNExecute(ctx)
 	}
-	return i.queryExecute(ec)
+	return i.queryExecute(ctx)
 }
 
-func (i *localIndexScan) topNExecute(ec executor.MeasureExecutionContext) (mit executor.MIterator, err error) {
+func (i *localIndexScan) topNExecute(ctx context.Context) (mit executor.MIterator, err error) {
 	var seriesList tsdb.SeriesList
+	ec := executor.FromMeasureExecutionContext(ctx)
 	for _, e := range i.entities {
 		shards, errInternal := ec.CompanionShards(i.metadata)
 		if errInternal != nil {
 			return nil, errInternal
 		}
 		for _, shard := range shards {
-			sl, errInternal := shard.Series().List(context.WithValue(
-				context.Background(),
-				logger.ContextKey,
-				i.l,
-			), tsdb.NewPath(e))
+			sl, errInternal := shard.Series().List(
+				context.WithValue(
+					ctx,
+					logger.ContextKey,
+					i.l,
+				), tsdb.NewPath(e))
 			if errInternal != nil {
 				return nil, errInternal
 			}
@@ -178,7 +181,7 @@ func (i *localIndexScan) topNExecute(ec executor.MeasureExecutionContext) (mit e
 		})
 	}
 	// CAVEAT: the order of series list matters when sorting by an index.
-	iters, closers, err := logical.ExecuteForShard(i.l, seriesList, i.timeRange, builders...)
+	iters, closers, err := logical.ExecuteForShard(ctx, i.l, seriesList, i.timeRange, builders...)
 	if err != nil {
 		return nil, err
 	}
@@ -193,19 +196,19 @@ func (i *localIndexScan) topNExecute(ec executor.MeasureExecutionContext) (mit e
 	if len(iters) == 0 {
 		return dummyIter, nil
 	}
-	tc := transformContext{
+	transformContext := transformContext{
 		ec:                   ec,
 		projectionTagsRefs:   i.projectionTagsRefs,
 		projectionFieldsRefs: i.projectionFieldsRefs,
 	}
 	if i.groupByEntity {
-		return newSeriesMIterator(iters, tc, i.maxDataPointsSize), nil
+		return newSeriesMIterator(iters, transformContext, i.maxDataPointsSize), nil
 	}
 	it := logical.NewItemIter(iters, i.order.Sort)
-	return newIndexScanIterator(it, tc, i.maxDataPointsSize), nil
+	return newIndexScanIterator(it, transformContext, i.maxDataPointsSize), nil
 }
 
-func (i *localIndexScan) queryExecute(ec executor.MeasureExecutionContext) (mit executor.MIterator, err error) {
+func (i *localIndexScan) queryExecute(ctx context.Context) (mit executor.MIterator, err error) {
 	var orderBy *tsdb.OrderBy
 	if i.order.Index != nil {
 		orderBy = &tsdb.OrderBy{
@@ -213,6 +216,7 @@ func (i *localIndexScan) queryExecute(ec executor.MeasureExecutionContext) (mit 
 			Sort:  i.order.Sort,
 		}
 	}
+	ec := executor.FromMeasureExecutionContext(ctx)
 	var seriesList tsdb.SeriesList
 	for _, e := range i.entities {
 		shards, errInternal := ec.Shards(e)
@@ -220,9 +224,9 @@ func (i *localIndexScan) queryExecute(ec executor.MeasureExecutionContext) (mit 
 			return nil, errInternal
 		}
 		for _, shard := range shards {
-			sl, errList := shard.Series().Search(
+			sl, errInternal := shard.Series().Search(
 				context.WithValue(
-					context.Background(),
+					ctx,
 					logger.ContextKey,
 					i.l,
 				),
@@ -230,8 +234,8 @@ func (i *localIndexScan) queryExecute(ec executor.MeasureExecutionContext) (mit 
 				i.filter,
 				orderBy,
 			)
-			if errList != nil {
-				return nil, errList
+			if errInternal != nil {
+				return nil, errInternal
 			}
 			seriesList = seriesList.Merge(sl)
 		}
@@ -246,7 +250,7 @@ func (i *localIndexScan) queryExecute(ec executor.MeasureExecutionContext) (mit 
 		})
 	}
 	// CAVEAT: the order of series list matters when sorting by an index.
-	iters, closers, err := logical.ExecuteForShard(i.l, seriesList, i.timeRange, builders...)
+	iters, closers, err := logical.ExecuteForShard(ctx, i.l, seriesList, i.timeRange, builders...)
 	if err != nil {
 		return nil, err
 	}
@@ -261,16 +265,16 @@ func (i *localIndexScan) queryExecute(ec executor.MeasureExecutionContext) (mit 
 	if len(iters) == 0 {
 		return dummyIter, nil
 	}
-	tc := transformContext{
+	transformContext := transformContext{
 		ec:                   ec,
 		projectionTagsRefs:   i.projectionTagsRefs,
 		projectionFieldsRefs: i.projectionFieldsRefs,
 	}
 	if i.groupByEntity {
-		return newSeriesMIterator(iters, tc, i.maxDataPointsSize), nil
+		return newSeriesMIterator(iters, transformContext, i.maxDataPointsSize), nil
 	}
 	it := logical.NewItemIter(iters, i.order.Sort)
-	return newIndexScanIterator(it, tc, i.maxDataPointsSize), nil
+	return newIndexScanIterator(it, transformContext, i.maxDataPointsSize), nil
 }
 
 func (i *localIndexScan) String() string {
@@ -309,7 +313,7 @@ func indexScan(startTime, endTime time.Time, metadata *commonv1.Metadata, projec
 var _ executor.MIterator = (*indexScanIterator)(nil)
 
 type indexScanIterator struct {
-	inner   logical.ItemIterator
+	inner   sort.Iterator[tsdb.Item]
 	err     error
 	current *measurev1.DataPoint
 	context transformContext
@@ -317,7 +321,7 @@ type indexScanIterator struct {
 	num     int
 }
 
-func newIndexScanIterator(inner logical.ItemIterator, context transformContext, max int) executor.MIterator {
+func newIndexScanIterator(inner sort.Iterator[tsdb.Item], context transformContext, max int) executor.MIterator {
 	return &indexScanIterator{
 		inner:   inner,
 		context: context,
@@ -326,10 +330,10 @@ func newIndexScanIterator(inner logical.ItemIterator, context transformContext, 
 }
 
 func (ism *indexScanIterator) Next() bool {
-	if !ism.inner.HasNext() || ism.err != nil || ism.num > ism.max {
+	if !ism.inner.Next() || ism.err != nil || ism.num > ism.max {
 		return false
 	}
-	nextItem := ism.inner.Next()
+	nextItem := ism.inner.Val()
 	var err error
 	if ism.current, err = transform(nextItem, ism.context); err != nil {
 		ism.err = multierr.Append(ism.err, err)