+++ CHANGES.md
@@ -14,6 +14,7 @@ Release Notes.
 - Register the node role to the metadata registry.
 - Implement the remote queue to spreading data to data nodes.
 - Fix parse environment variables error
+- Implement the distributed query engine.
 
 ### Bugs